+++ mlpf/met_finetuning_pipeline.py
@@ -15,8 +15,8 @@ import torch.distributed as dist
 import torch.multiprocessing as mp
 import torch.nn as nn
 import yaml
-from pyg.finetuning import configure_model_trainable, finetune_mlpf, override_config
 from pyg.logger import _configLogger, _logger
+from pyg.met_finetuning import configure_model_trainable, finetune_mlpf, override_config
 from pyg.mlpf import MLPF
 from pyg.PFDataset import get_interleaved_dataloaders
 from pyg.utils import count_parameters, load_checkpoint, save_HPs
